move dumps to ee
