+++ images/agent/src/internal/utils/thin_dump_ee_test.go
@@ -1,3 +1,10 @@
+//go:build !ce
+
+/*
+Copyright 2025 Flant JSC
+Licensed under the Deckhouse Platform Enterprise Edition (EE) license. See https://github.com/deckhouse/deckhouse/blob/main/ee/LICENSE
+*/
+
 package utils_test
 
 import (
